+++ src/extension.ts
@@ -4,6 +4,7 @@ import * as path from 'path';
 import * as fs from 'fs';
 import {spawn} from 'child_process';
 import * as moment from 'moment';
+import * as upath from 'upath';
 
 export function activate(context: vscode.ExtensionContext) {
     console.log('Congratulations, your extension "vscode-paste-image" is now active!');
@@ -19,6 +20,16 @@ export function deactivate() {
 }
 
 class Paster {
+    static PATH_VARIABLE_CURRNET_FILE_DIR = /\$\{currentFileDir\}/;
+	static PATH_VARIABLE_PROJECT_ROOT = /\$\{projectRoot\}/;
+    static PATH_VARIABLE_CURRNET_FILE_NAME = /\$\{currentFileName\}/;
+    static PATH_VARIABLE_CURRNET_FILE_NAME_WITHOUT_EXT = /\$\{currentFileNameWithoutExt\}/;
+
+    static folderPathFromConfig:string;
+    static basePathFromConfig:string;
+    static prefixFromConfig:string;
+    static suffixFromConfig:string;
+    static forceUnixStyleSeparatorFromConfig:boolean;
 
     public static paste() {
         // get current edit file path
@@ -31,34 +42,69 @@ class Paster {
             vscode.window.showInformationMessage('Before paste image, you need to save current edit file first.');
             return;
         }
+        let filePath = fileUri.fsPath;
+        let folderPath = path.dirname(filePath);
+        let projectPath = vscode.workspace.rootPath;
 
         // get selection as image file name, need check
         var selection = editor.selection;
         var selectText = editor.document.getText(selection);
-        if(selectText && !/^[\w\-.]+$/.test(selectText)){
+        if(selectText && !/^[\w\-.\/]+$/.test(selectText)){
             vscode.window.showInformationMessage('Your selection is not a valid file name!');
             return;
         }
 
-        // get image destination path
-        let folderPathFromConfig = vscode.workspace.getConfiguration('pasteImage')['path'];
-        if (folderPathFromConfig && (folderPathFromConfig.length !== folderPathFromConfig.trim().length)) {
-            vscode.window.showErrorMessage('The specified path is invalid. "' + folderPathFromConfig + '",please check your config.');
+        // load config pasteImage.path/pasteImage.basePath
+        this.folderPathFromConfig = vscode.workspace.getConfiguration('pasteImage')['path'];
+        if(!this.folderPathFromConfig){
+            this.folderPathFromConfig = "${currentFileDir}";
+        }
+        if (this.folderPathFromConfig.length !== this.folderPathFromConfig.trim().length) {
+            vscode.window.showErrorMessage(`The config pasteImage.path = '${this.folderPathFromConfig}' is invalid. please check your config.`);
             return;
         }
-        let filePath = fileUri.fsPath;
-        let imagePath = this.getImagePath(filePath, selectText, folderPathFromConfig);
+        this.basePathFromConfig = vscode.workspace.getConfiguration('pasteImage')['basePath'];
+        if(!this.basePathFromConfig){
+            this.basePathFromConfig = "";
+        }
+        if (this.basePathFromConfig.length !== this.basePathFromConfig.trim().length) {
+            vscode.window.showErrorMessage(`The config pasteImage.path = '${this.basePathFromConfig}' is invalid. please check your config.`);
+            return;
+        }
+        this.prefixFromConfig = vscode.workspace.getConfiguration('pasteImage')['prefix'];
+        this.suffixFromConfig = vscode.workspace.getConfiguration('pasteImage')['suffix'];
+        this.forceUnixStyleSeparatorFromConfig = vscode.workspace.getConfiguration('pasteImage')['forceUnixStyleSeparator'];
+        this.forceUnixStyleSeparatorFromConfig = !!this.forceUnixStyleSeparatorFromConfig;
 
+        this.folderPathFromConfig = this.replacePathVariable(this.folderPathFromConfig, projectPath, filePath);
+        this.basePathFromConfig = this.replacePathVariable(this.basePathFromConfig, projectPath, filePath);
+
+        let imagePath = this.getImagePath(filePath, selectText, this.folderPathFromConfig);
+
+        // is the file existed?
+        let existed = fs.existsSync(imagePath);
+        if(existed){
+            vscode.window.showInformationMessage(`File ${imagePath} existed.Would you want to replace?`,'Replace','Cancel').then(choose =>{
+                if(choose != 'Replace') return; 
+
+                this.saveAndPaste(editor,imagePath);
+            });
+        }else{
+            this.saveAndPaste(editor,imagePath);
+        }
+    }
+
+    public static saveAndPaste(editor: vscode.TextEditor, imagePath){
         this.createImageDirWithImagePath(imagePath).then(imagePath => {
             // save image and insert to current edit file
-            this.saveClipboardImageToFileAndGetPath(imagePath, imagePath => {
-                if(!imagePath) return;
-                if(imagePath === 'no image'){
+            this.saveClipboardImageToFileAndGetPath(imagePath, (imagePath,imagePathReturnByScript) => {
+                if(!imagePathReturnByScript) return;
+                if(imagePathReturnByScript === 'no image'){
                     vscode.window.showInformationMessage('There is not a image in clipboard.');
                     return;
                 }
 
-                imagePath = this.renderFilePath(editor.document.languageId,filePath,imagePath);
+                imagePath = this.renderFilePath(editor.document.languageId,this.basePathFromConfig,imagePath,this.forceUnixStyleSeparatorFromConfig,this.prefixFromConfig,this.suffixFromConfig);
 
                 editor.edit(edit => {
                     let current = editor.selection;
@@ -126,7 +172,7 @@ class Paster {
     /**
      * use applescript to save image from clipboard and get file path
      */
-    private static saveClipboardImageToFileAndGetPath(imagePath,cb:(imagePath:string)=>void) {
+    private static saveClipboardImageToFileAndGetPath(imagePath,cb:(imagePath:string,imagePathFromScript:string)=>void) {
         if (!imagePath) return;
 
         let platform = process.platform;
@@ -147,7 +193,7 @@ class Paster {
                 // console.log('exit', code, signal);
             });
             powershell.stdout.on('data', function (data: Buffer) {
-                cb(data.toString().trim());
+                cb(imagePath,data.toString().trim());
             });
         }
         else if(platform === 'darwin'){
@@ -160,7 +206,7 @@ class Paster {
             });
 
             ascript.stdout.on('data', function (data:Buffer) {
-                cb(data.toString().trim());
+                cb(imagePath,data.toString().trim());
             });
         } else {
             // Linux 
@@ -178,7 +224,7 @@ class Paster {
                     vscode.window.showInformationMessage('You need to install xclip command first.');
                     return;
                 }
-                cb(result);
+                cb(imagePath,result);
             });
         }
     }
@@ -187,8 +233,16 @@ class Paster {
      * render the image file path dependen on file type
      * e.g. in markdown image file path will render to ![](path)
      */
-    public static renderFilePath(languageId:string,docPath:string,imageFilePath:string):string{
-        imageFilePath = path.relative(path.dirname(docPath),imageFilePath);
+    public static renderFilePath(languageId:string,basePath:string,imageFilePath:string,forceUnixStyleSeparator:boolean,prefix:string,suffix:string):string{
+        if(basePath){
+            imageFilePath = path.relative(basePath,imageFilePath);
+        }
+
+        if(forceUnixStyleSeparator){
+            imageFilePath = upath.normalize(imageFilePath);
+        }
+
+        imageFilePath = `${prefix}${imageFilePath}${suffix}`;
 
         switch(languageId){
             case "markdown":
@@ -198,5 +252,18 @@ class Paster {
             default:
                 return imageFilePath;
         }
+    }
+
+    public static replacePathVariable(pathStr:string,projectRoot:string,curFilePath:string):string{
+        let currentFileDir = path.dirname(curFilePath);
+        let ext = path.extname(curFilePath);
+        let fileName = path.basename(curFilePath);
+        let fileNameWithoutExt = path.basename(curFilePath,ext);
+
+        pathStr = pathStr.replace(this.PATH_VARIABLE_PROJECT_ROOT, projectRoot);
+        pathStr = pathStr.replace(this.PATH_VARIABLE_CURRNET_FILE_DIR, currentFileDir);
+        pathStr = pathStr.replace(this.PATH_VARIABLE_CURRNET_FILE_NAME, fileName);
+        pathStr = pathStr.replace(this.PATH_VARIABLE_CURRNET_FILE_NAME_WITHOUT_EXT, fileNameWithoutExt);
+        return pathStr;
     }
 }