+++ typings/upath.d.ts
@@ -1,0 +1,3 @@
+declare module 'upath'{
+    function normalize(input:string):string;
+}